+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## v1.4.1
+
+### Important Notes
+
+- Team users update route using the id instead of the username.
+- Optimize check permisiion logic. 
+- Add update admin route.
+
+### Potential bug fix
+
+- Fix: golang.org/x/net vulnerable to Cross-site Scripting
+
 ## v1.4.0
 
 ### Features